Tc apidocs 841 updates (#35)
* Update advertiser-dataprovider-guide.md

Added a note and an FAQ on the number of buckets, etc.

* Update advertiser-dataprovider-guide.md

Added a link in an FAQ.

* Update advertiser-dataprovider-guide.md

Misc. wording updates.

* Update advertiser-dataprovider-guide.md

Added an FAQ on handling optouts.

* Update get-token-validate.md

Added a note on URL encoding in the token query parameter.

* Update get-token-refresh.md

Minor wording edits.

* Update advertiser-dataprovider-guide.md

fixed a grammar issue.

* Update advertiser-dataprovider-guide.md

Fixed typos and updated the optout FAQ wording.

* Update advertiser-dataprovider-guide.md

Added a link to the portal.

* Update get-token-refresh.md

fixed a grammar issue.